refactor(mark): consolidate service implementation and introduce golden example
- Convert `MarkService` from an interface to a concrete class and remove `MarkServiceImpl`
- Delete unused `MarkOccurrenceRequestDto` and `MarkRequestDto` records
- Add `goldenExample` relation to `Mark` entity and support updates via `MarkUpdateDto`
- Update `MarkMapper` and `MarkDto` to extract `coverId` (now UUID) and `embedding` from the golden example
- Map `createdAt` correctly in `MarkOccurrenceMapper` instead of returning null
