+++ mark/src/main/java/pt/estga/mark/dtos/MarkDto.java
@@ -1,10 +1,12 @@
 package pt.estga.mark.dtos;
 
+import java.util.UUID;
+
 public record MarkDto(
         Long id,
         String title,
         String description,
         float[] embedding,
-        Long coverId,
+        UUID coverId,
         Boolean active
 ) { }
