--- mark/src/main/java/pt/estga/mark/dtos/MarkRequestDto.java
@@ -1,7 +1,0 @@
-package pt.estga.mark.dtos;
-
-public record MarkRequestDto(
-        String description,
-        Long coverId,
-        Boolean active
-) { }
