+++ mark/src/main/java/pt/estga/mark/dtos/MarkUpdateDto.java
@@ -1,7 +1,10 @@
 package pt.estga.mark.dtos;
 
+import java.util.UUID;
+
 public record MarkUpdateDto(
         String title,
         String description,
+        UUID goldenExampleId,
         Boolean active
 ) { }
