+++ mark/src/main/java/pt/estga/mark/mappers/MarkMapper.java
@@ -2,6 +2,7 @@ package pt.estga.mark.mappers;
 
 import pt.estga.mark.dtos.MarkDto;
 import pt.estga.mark.entities.Mark;
+import pt.estga.mark.entities.MarkEvidence;
 import pt.estga.shared.enums.EntityStatus;
 
 public class MarkMapper {
@@ -10,12 +11,13 @@ public class MarkMapper {
 
     public static MarkDto toDto(Mark mark) {
         if (mark == null) return null;
+        MarkEvidence golden = mark.getGoldenExample();
         return new MarkDto(
                 mark.getId(),
                 mark.getTitle(),
                 mark.getDescription(),
-                null,
-                null,
+                golden != null ? golden.getEmbedding() : null,
+                golden != null ? golden.getFileId() : null,
                 mark.getStatus() == EntityStatus.ACTIVE
         );
     }
